third_party/blink/web_tests/xr/xrWebGLLayer_constructor.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/testdriver.js"></script>
<script src="../resources/testdriver-vendor.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../external/wpt/webxr/resources/webxr_test_constants.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<canvas />

<script>
let testName =
  "Ensure a WebGL layer's framebuffer can only be drawn to inside a XR frame";

let fakeDeviceInitParams = TRACKED_IMMERSIVE_DEVICE;

let testFunction =
  (session, fakeDeviceController, t, sessionObjects) => new Promise((resolve, reject) => {
  let gl = sessionObjects.gl;
  let webglCanvas = document.getElementsByTagName('canvas')[0];

  try {
    let webglLayerGood = new XRWebGLLayer(session, gl);
  } catch (err) {
    reject("XRWebGLLayer should not fail with valid arguments");
  }

  let lose_context_ext = gl.getExtension('WEBGL_lose_context');

  webglCanvas.addEventListener('webglcontextlost', (ev) => {
    ev.preventDefault();

    try {
      let webglLayerBadContext = new XRWebGLLayer(session, gl);
      reject("XRWebGLLayer should fail when created with a lost context");
    } catch (err) {
      assert_equals(err.name, 'InvalidStateError');
      setTimeout(() => { lose_context_ext.restoreContext(); }, 100);
    }
  });

  webglCanvas.addEventListener('webglcontextrestored', (ev) => {
    resolve(session.end().then(() => {
      try {
        let webglLayerBadSession = new XRWebGLLayer(session, gl);
        assert_unreached("XRWebGLLayer should fail when created with an ended session");
      } catch (err) {
        assert_equals(err.name, 'InvalidStateError');
      }
    }));
  });

  lose_context_ext.loseContext();

});

xr_session_promise_test(
  testName, testFunction, fakeDeviceInitParams, 'immersive-vr');

</script>
